==> literature review/2_code/2_label_sample.ipynb <==
import pandas as pd
import ipywidgets as widgets
from IPython.display import display, clear_output, HTML
import textwrap
from functools import partial
import pandas as pd

# Read the sample data
df_sample = pd.read_csv('../1_data/sample1000_papers.csv')
df_sample.head(2)

# Crear un botón para confirmar la selección
confirm_button = widgets.Button(
    description='confirm',
    disabled=False,
    button_style='',  # 'success', 'info', 'warning', 'danger' o ''
    tooltip='Confirm selection',
    icon='check'
)

# Función para mostrar la siguiente observación
def display_next_observation(index, df, select_box):
    if index < 401:
        clear_output(wait=True)
        print(f"Index: {index}")
        print("-----------------------------")
        print(f"Title: {df.iloc[index]['Title']}")
        print("-----------------------------")
        print("Abstract:")
        print(textwrap.fill(df.iloc[index]['Abstract'], width=150))
        print("-----------------------------")
        print(f"Keywords: {df.iloc[index]['Author Keywords']}")
        display(select_box)
        display(confirm_button)
    else:
        clear_output(wait=True)
        print("200 papers have been reviewed.")

# Función para manejar el clic del botón
def on_button_click(current_index, df, select_box, button):
    df.at[df.index[current_index[0]], 'selected'] = select_box.value
    current_index[0] += 1
    display_next_observation(current_index[0], df, select_box)

select_box = widgets.Select(
    options=['yes', 'no'],
    value=None,
    description='Select:',
    disabled=False
)

# Inicializar el índice actual
current_index = [min(df_sample.loc[pd.isna(df_sample['selected'])].index)]

# Conectar el evento del botón a la función
confirm_button.on_click(partial(on_button_click, current_index, df_sample, select_box))

# Mostrar la primera observación
display(HTML("<style>.container { width:100% !important; }</style>"))
display_next_observation(current_index[0], df_sample, select_box)


# Count the instances of selected
df_sample["selected"].value_counts()

# Mark labeled data
df_sample['human_labeled'] = df_sample['selected'].apply(lambda x: 0 if pd.isna(x) else 1)
print(df_sample["human_labeled"].value_counts())
df_sample

# Save the data
df_sample.to_csv('../1_data/sample1000_papers.csv', index=False)
